Reorganized the packaging structure
